requirements.txt
allure_python_commons==2.13.1
jsonpath==0.82
pytest==7.2.2
PyYAML==6.0
PyYAML==6.0
Requests==2.31.0
requests_toolbelt==1.0.0
